# Conflicts:
#	src/main/web/src/app/shared/services/comment.service.ts
#	src/main/web/src/app/shared/services/user.service.ts

+++ src/main/web/src/app/shared/services/participants.service.ts
@@ -75,6 +75,11 @@ export class ParticipantsService {
 	addOrModify(requestMethod: (url: string, body: any, options?: RequestOptionsArgs) => Observable<Response>,
 				eventId: number, eventType: EventType, participant: Participant): Observable<any> {
 
+		//todo remove demo
+		if (eventId >= 0) {
+			return Observable.of(participant);
+		}
+
 		return requestMethod("/api/participants", {eventId, eventType, participant})
 			.map(response => response.json() as any)
 			//retry 3 times before throwing an error
@@ -114,6 +119,11 @@ export class ParticipantsService {
 	 * @param participantId
 	 */
 	deleteParticipant(eventId: number, eventType: EventType, participantId: number): Observable<any> {
+		//todo remove demo
+		if (eventId >= 0) {
+			return Observable.of(null);
+		}
+
 		return this.http.delete("/api/participants", {body: {eventId, eventType, participantId}})
 			.map(response => response.json() as any)
 			//retry 3 times before throwing an error
